refactor(core): fold rss-api into core-api (ADR-0002, final fold) [REL-16]
The last and only fold with real writes. Core now owns the feed-target
tables (tracked_feeds, user_custom_feeds) the Rust poller reads, and
runs the auto-cleanup janitor. Route shapes, cache keys, and bodies are
identical; the desktop app needs no changes.

- api/core/rss.go: /rss/feeds GET (per-user catalog, singleflight) +
  DELETE (multi-tenant custom-feed removal, transaction intact) +
  /rss/health; dashboard section; syncRSSFeedsToTracked moved verbatim
- Janitor: same thresholds/cadence/transaction shape, now guarded by a
  Redis SET NX lock (lock:rss-janitor, 10m TTL) since core runs 2
  replicas — one sweeps, the other skips. Started from api/main.go next
  to the other background workers.
- localSource gains an optional full `lifecycle` hook — rss is the one
  source whose lifecycle does real work (feed sync) beyond cache
  invalidation; the other three keep invalidateUser only
- Auth: catalog/delete derive the user from LogtoAuth (GetUserID), not
  the proxy-injected X-User-Sub header
- Dropped per REL-10: /internal/cdc handler and all rss:subscribers:*
  writes (the sets have a 7-day TTL and now zero writers — they age out)
- Removed: channels/rss/api module, k8s/rss-api.yaml, its NetworkPolicy
  pair (rss-service now admits core-api), deploy matrix row + filter,
  backend-tests entry, smoke entries, compose blocks, legacy channel
  compose files; INTERNAL_RSS_URL moved to core-config
- Docs: AGENTS.md, LOCAL_SETUP.md — fantasy is now the one remaining
  channel Go API; discovery serves a single registrant

Post-merge manual cleanup (kubectl apply does not prune):
  kubectl delete deployment/rss-api service/rss-api -n scrollr
  kubectl delete networkpolicy allow-core-to-rss-api allow-rss-api-to-rss-service -n scrollr
