+++ .github/workflows/deploy.yml
@@ -38,7 +38,6 @@ jobs:
       website: ${{ steps.manual.outputs.website || steps.changes.outputs.website }}
       finance-service: ${{ steps.manual.outputs.finance-service || steps.changes.outputs.finance-service }}
       sports-service: ${{ steps.manual.outputs.sports-service || steps.changes.outputs.sports-service }}
-      rss-api: ${{ steps.manual.outputs.rss-api || steps.changes.outputs.rss-api }}
       rss-service: ${{ steps.manual.outputs.rss-service || steps.changes.outputs.rss-service }}
       fantasy-api: ${{ steps.manual.outputs.fantasy-api || steps.changes.outputs.fantasy-api }}
       predictions-service: ${{ steps.manual.outputs.predictions-service || steps.changes.outputs.predictions-service }}
@@ -55,7 +54,7 @@ jobs:
           # website, and only for `desktop-v*` tags. Other releases (none
           # exist today, but future-proof) are ignored.
           if [ "${{ github.event_name }}" = "release" ]; then
-            for svc in core-api website finance-service sports-service rss-api rss-service fantasy-api predictions-service k8s; do
+            for svc in core-api website finance-service sports-service rss-service fantasy-api predictions-service k8s; do
               echo "${svc}=false" >> $GITHUB_OUTPUT
             done
 
@@ -75,14 +74,14 @@ jobs:
             exit 0
           fi
 
-          for svc in core-api website finance-service sports-service rss-api rss-service fantasy-api predictions-service k8s; do
+          for svc in core-api website finance-service sports-service rss-service fantasy-api predictions-service k8s; do
             echo "${svc}=false" >> $GITHUB_OUTPUT
           done
 
           services="$(printf '%s' '${{ inputs.services }}' | tr '[:upper:]' '[:lower:]' | tr -d '[:space:]')"
 
           if [ -z "$services" ] || [ "$services" = "all" ]; then
-            for svc in core-api website finance-service sports-service rss-api rss-service fantasy-api predictions-service k8s; do
+            for svc in core-api website finance-service sports-service rss-service fantasy-api predictions-service k8s; do
               echo "${svc}=true" >> $GITHUB_OUTPUT
             done
             exit 0
@@ -91,7 +90,7 @@ jobs:
           IFS=',' read -r -a selected <<< "$services"
           for svc in "${selected[@]}"; do
             case "$svc" in
-              core-api|website|finance-service|sports-service|rss-api|rss-service|fantasy-api|predictions-service|k8s)
+              core-api|website|finance-service|sports-service|rss-service|fantasy-api|predictions-service|k8s)
                 echo "${svc}=true" >> $GITHUB_OUTPUT
                 ;;
               *)
@@ -114,8 +113,6 @@ jobs:
               - 'channels/finance/service/**'
             sports-service:
               - 'channels/sports/service/**'
-            rss-api:
-              - 'channels/rss/api/**'
             rss-service:
               - 'channels/rss/service/**'
             fantasy-api:
@@ -149,10 +146,6 @@ jobs:
             context: ./channels/sports/service
             dockerfile: ./channels/sports/service/Dockerfile
             changed: ${{ needs.detect-changes.outputs.sports-service }}
-          - service: rss-api
-            context: ./channels/rss/api
-            dockerfile: ./channels/rss/api/Dockerfile
-            changed: ${{ needs.detect-changes.outputs.rss-api }}
           - service: rss-service
             context: ./channels/rss/service
             dockerfile: ./channels/rss/service/Dockerfile
@@ -331,7 +324,6 @@ jobs:
           kubectl apply -f k8s/sports-service.yaml
           kubectl apply -f k8s/rss-service.yaml
           kubectl apply -f k8s/predictions-service.yaml
-          kubectl apply -f k8s/rss-api.yaml
           kubectl apply -f k8s/fantasy-api.yaml
           kubectl apply -f k8s/core-api.yaml
           kubectl apply -f k8s/website.yaml
@@ -434,7 +426,6 @@ jobs:
           rollout_if_needed website "${{ needs.detect-changes.outputs.website }}"
           rollout_if_needed finance-service "${{ needs.detect-changes.outputs.finance-service }}"
           rollout_if_needed sports-service "${{ needs.detect-changes.outputs.sports-service }}"
-          rollout_if_needed rss-api "${{ needs.detect-changes.outputs.rss-api }}"
           rollout_if_needed rss-service "${{ needs.detect-changes.outputs.rss-service }}"
           rollout_if_needed fantasy-api "${{ needs.detect-changes.outputs.fantasy-api }}"
           rollout_if_needed predictions-service "${{ needs.detect-changes.outputs.predictions-service }}"
