--- channels/rss/api/go.mod
@@ -1,34 +1,0 @@
-module github.com/brandon-relentnet/scrollr-rss
-
-go 1.25.0
-
-require (
-	github.com/getsentry/sentry-go v0.46.2
-	github.com/getsentry/sentry-go/fiber v0.46.2
-	github.com/gofiber/fiber/v2 v2.52.12
-	github.com/jackc/pgx/v5 v5.5.3
-	github.com/joho/godotenv v1.5.1
-	github.com/redis/go-redis/v9 v9.4.0
-	golang.org/x/sync v0.8.0
-)
-
-require (
-	github.com/andybalholm/brotli v1.1.1 // indirect
-	github.com/cespare/xxhash/v2 v2.2.0 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/google/uuid v1.6.0 // indirect
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
-	github.com/jackc/puddle/v2 v2.2.1 // indirect
-	github.com/klauspost/compress v1.17.11 // indirect
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	github.com/mattn/go-runewidth v0.0.16 // indirect
-	github.com/rivo/uniseg v0.4.7 // indirect
-	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasthttp v1.57.0 // indirect
-	github.com/valyala/tcplisten v1.0.0 // indirect
-	golang.org/x/crypto v0.28.0 // indirect
-	golang.org/x/sys v0.28.0 // indirect
-	golang.org/x/text v0.19.0 // indirect
-)
