--- channels/rss/docker-compose.dev.yml
@@ -1,25 +1,0 @@
-version: "3.8"
-
-services:
-  scrollr-rss-api:
-    build:
-      context: ./api
-      dockerfile: Dockerfile
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - REDIS_URL=${REDIS_URL}
-      - ALLOWED_ORIGINS=${ALLOWED_ORIGINS}
-      - CHANNEL_URL=https://rss.api.myscrollr.enanimate.dev
-      - INTERNAL_RSS_URL=http://scrollr-rss-service:3004
-    depends_on:
-      scrollr-rss-service:
-        condition: service_started
-    restart: unless-stopped
-
-  scrollr-rss-service:
-    build:
-      context: ./service
-      dockerfile: Dockerfile
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-    restart: unless-stopped
