--- channels/rss/docker-compose.yml
@@ -1,34 +1,0 @@
-version: "3.8"
-
-services:
-  scrollr-rss-api:
-    build:
-      context: ./api
-      dockerfile: Dockerfile
-    container_name: scrollr-rss-api
-    # Bind only to localhost — the core gateway is the only trusted caller
-    # and reaches us via the Docker network (CHANNEL_URL). Publishing on
-    # 0.0.0.0 let anything else on the host bypass the gateway and hit
-    # our routes directly (channels don't validate JWTs — they trust the
-    # X-User-Sub header set by the gateway).
-    ports:
-      - "127.0.0.1:8083:8083"
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-      - REDIS_URL=${REDIS_URL}
-      - ENCRYPTION_KEY=${ENCRYPTION_KEY}
-      - ALLOWED_ORIGINS=${ALLOWED_ORIGINS}
-      - CHANNEL_URL=${CHANNEL_URL}
-      - INTERNAL_RSS_URL=http://scrollr-rss-service:3004
-    restart: unless-stopped
-
-  scrollr-rss-service:
-    build:
-      context: ./service
-      dockerfile: Dockerfile
-    container_name: scrollr-rss-service
-    ports:
-      - "3004:3004"
-    environment:
-      - DATABASE_URL=${DATABASE_URL}
-    restart: unless-stopped
