+++ k8s/configmap-core.yaml
@@ -24,6 +24,7 @@ data:
   # ADR-0002). Moved from channels-config as each Go channel API folds in.
   INTERNAL_FINANCE_URL: "http://finance-service:3001"
   INTERNAL_SPORTS_URL: "http://sports-service:3002"
+  INTERNAL_RSS_URL: "http://rss-service:3004"
   INTERNAL_PREDICTIONS_URL: "http://predictions-service:3005"
 
   # Logto role IDs (not secret, just config)
