--- k8s/rss-api.yaml
@@ -1,125 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: rss-api
-  namespace: scrollr
-  labels:
-    app: rss-api
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: rss-api
-  template:
-    metadata:
-      labels:
-        app: rss-api
-    spec:
-      containers:
-        - name: rss-api
-          image: registry.digitalocean.com/scrollr/rss-api:latest
-          ports:
-            - containerPort: 8083
-          env:
-            - name: CHANNEL_URL
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: RSS_CHANNEL_URL
-            - name: INTERNAL_RSS_URL
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: INTERNAL_RSS_URL
-            - name: ALLOWED_ORIGINS
-              valueFrom:
-                configMapKeyRef:
-                  name: channels-config
-                  key: ALLOWED_ORIGINS
-            - name: DATABASE_URL
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: DATABASE_URL
-            - name: REDIS_URL
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: REDIS_URL
-            - name: ENCRYPTION_KEY
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: ENCRYPTION_KEY
-            # Sentry — per-service DSN inlined (ingestion-only, not a leak).
-            - name: SENTRY_DSN
-              value: "https://341add4a3b2cb8d51ae93cfdb83832eb@o4511384091033600.ingest.us.sentry.io/4511384401215489"
-            - name: SENTRY_USER_SALT
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: SENTRY_USER_SALT
-                  optional: true
-            - name: ENVIRONMENT
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: ENVIRONMENT
-                  optional: true
-            - name: GIT_SHA
-              valueFrom:
-                secretKeyRef:
-                  name: scrollr-secrets
-                  key: GIT_SHA
-                  optional: true
-          resources:
-            requests:
-              cpu: 50m
-              memory: 64Mi
-            limits:
-              cpu: 250m
-              memory: 128Mi
-          # /internal/health now pings DB, Redis, and the Rust ingestion
-          # service's /health/ready (see PR #108). It returns 503 when
-          # any dependency is down.
-          startupProbe:
-            httpGet:
-              path: /internal/health
-              port: 8083
-            initialDelaySeconds: 3
-            periodSeconds: 5
-            failureThreshold: 30
-            timeoutSeconds: 3
-          livenessProbe:
-            httpGet:
-              path: /internal/health
-              port: 8083
-            initialDelaySeconds: 10
-            periodSeconds: 30
-            failureThreshold: 6
-            timeoutSeconds: 3
-          readinessProbe:
-            httpGet:
-              path: /internal/health
-              port: 8083
-            initialDelaySeconds: 5
-            periodSeconds: 10
-            failureThreshold: 3
-            timeoutSeconds: 3
-          lifecycle:
-            preStop:
-              exec:
-                command: ["/bin/sh", "-c", "sleep 10"]
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: rss-api
-  namespace: scrollr
-spec:
-  selector:
-    app: rss-api
-  ports:
-    - port: 8083
-      targetPort: 8083
-  type: ClusterIP
